Abstract common request handling parts, use inversify for differentiation between singe/multi user approach.
