+++ tests/e2e/utils/SingleUserRequestHandler.ts
@@ -1,0 +1,8 @@
+import { injectable } from 'inversify';
+import { IRequestHandler } from './IRequestHandler';
+import { AbstractRequestHandler } from './AbstractRequestHandler';
+
+@injectable()
+export class SingleUserRequestHandler extends AbstractRequestHandler implements IRequestHandler {
+
+}
